requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
whitenoise>=6.7,<7.0
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<24.0
